Move ProcessingAnimation to atoms
